requirements.txt
fastapi
uvicorn
pyqrcode
pypng